requirements.txt
boto3==1.38.21
# chromadb==1.0.13
datasets==3.6.0
python-docx
pinecone-client
opensearch-py
ebooklib
GitPython==3.1.44
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.2
html2text==2025.4.15
langchain==0.3.25
langchain-aws
# langchain-chroma==0.2.4
langchain-huggingface==0.2.0
langchain-openai==0.3.17
langchain-pinecone==0.2.6
langchain-community
langchain-tests==0.3.19
langchain-text-splitters==0.3.8
lxml==5.4.0
markdown-it-py==3.0.0
numpy
pandas==2.2.3
peft==0.15.2
pypdf==5.5.0
pymupdf
streamlit==1.45.1
tokenizers==0.21.1
torch==2.7.0
transformers==4.52.3